+++ frontend/src/components/Navbar/MeDropdown.tsx
@@ -5,7 +5,7 @@ import { FcInfo, FcQuestions, FcFeedback, FcPuzzle } from 'react-icons/fc';
 import { FaSignOutAlt, FaSignInAlt } from 'react-icons/fa';
 
 import styles from './MeDropdown.module.css';
-import { logout, scrollToTop } from '../../utilities';
+import { logout, scrollToTop } from '../../utilities/display';
 import {
   SurfaceComponent,
   TextComponent,
