+++ frontend/src/components/Search/NoCourses.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 import NoCoursesFound from '../../images/no_courses_found.svg';
-import { toSeasonString } from '../../utilities/courseUtilities';
+import { toSeasonString } from '../../utilities/course';
 import { useWorksheet } from '../../contexts/worksheetContext';
 
 function NoCourses() {
