+++ frontend/src/components/Worksheet/CalendarEvent.tsx
@@ -4,7 +4,7 @@ import LinesEllipsis from 'react-lines-ellipsis';
 import responsiveHOC from 'react-lines-ellipsis/lib/responsiveHOC';
 import { StyledPopover } from '../StyledComponents';
 import type { Listing } from '../../utilities/common';
-import { truncatedText } from '../../utilities/courseUtilities';
+import { truncatedText } from '../../utilities/course';
 
 const ResponsiveEllipsis = responsiveHOC()(LinesEllipsis);
