+++ frontend/src/utilities/common.ts
