+++ frontend/src/utilities/course.ts
